chore : MemberQueryService 분리
